refactor(os): merge ryvex-os into ryvex-term
same thing, makes more sense this way
