--- ryvex-os/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "ryvex-os"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-proc-macros = { path = "../proc-macros" }
