+++ ryvex-term/src/error.rs
@@ -8,12 +8,14 @@ use proc_macros::StackTraceDebug;
 #[derive(StackTraceDebug)]
 pub enum TermError {
 	IoError(std::io::Error),
+	TermiosError(std::io::Error),
 }
 
 impl Error for TermError {
 	fn source(&self) -> Option<&(dyn Error + 'static)> {
 		match self {
 			TermError::IoError(error) => Some(error),
+			TermError::TermiosError(error) => Some(error),
 		}
 	}
 
@@ -36,6 +38,7 @@ impl Display for TermError {
 	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
 		match self {
 			TermError::IoError(_) => write!(f, "IO error"),
+			TermError::TermiosError(_) => write!(f, "Termios error"),
 		}
 	}
 }
