+++ ryvex-term/src/lib.rs
@@ -1,3 +1,5 @@
 pub mod error;
 pub mod event;
 pub mod key;
+pub mod sys;
+pub mod termios;
