+++ ryvex-term/src/sys/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(unix)]
+pub mod unix;
+#[cfg(windows)]
+pub mod windows;
